Refactor record driver hierarchy (VUFIND-1240) (#1065)
