+++ module/VuFind/src/VuFind/RecordDriver/EIT.php
@@ -36,7 +36,7 @@ namespace VuFind\RecordDriver;
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org/wiki/development:plugins:record_drivers Wiki
  */
-class EIT extends SolrDefault
+class EIT extends DefaultRecord
 {
     /**
      * Used for identifying search backends
