+++ module/VuFind/src/VuFind/RecordDriver/LibGuides.php
@@ -36,7 +36,7 @@ namespace VuFind\RecordDriver;
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org/wiki/development:plugins:record_drivers Wiki
  */
-class LibGuides extends SolrDefault
+class LibGuides extends DefaultRecord
 {
     /**
      * Get the short (pre-subtitle) title of the record.
