+++ module/VuFind/src/VuFind/RecordDriver/Primo.php
@@ -36,7 +36,7 @@ namespace VuFind\RecordDriver;
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org/wiki/development:plugins:record_drivers Wiki
  */
-class Primo extends SolrDefault
+class Primo extends DefaultRecord
 {
     /**
      * Date converter
