+++ themes/bootstrap3/templates/RecordDriver/DefaultRecord/core-qrcode.phtml
