+++ themes/bootstrap3/templates/RecordDriver/DefaultRecord/format-list.phtml
