+++ themes/bootstrap3/templates/RecordDriver/DefaultRecord/link-journaltitle.phtml
