+++ themes/bootstrap3/templates/RecordDriver/DefaultRecord/list-entry.phtml
